% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_fill_Nas.R
\name{sr_fill_nas}
\alias{sr_fill_nas}
\title{Title sr_fill_nas
Cette fonction remplit les éventuels NA présents dans un jeu de données
de supports et résistances en utilisant la méthode Last Observation Carried Forward (LOCF).
Cela permet de maintenir la continuité des niveaux critiques dans l'analyse technique.}
\usage{
sr_fill_nas(sr_data)
}
\arguments{
\item{sr_data}{data.frame ou matrix numérique contenant les niveaux de supports et résistances,
dont les NA seront remplis colonne par colonne.}
}
\value{
Un objet du même type que \code{sr_data} avec les NA remplacés par la dernière valeur non manquante observée.
}
\description{
Title sr_fill_nas
Cette fonction remplit les éventuels NA présents dans un jeu de données
de supports et résistances en utilisant la méthode Last Observation Carried Forward (LOCF).
Cela permet de maintenir la continuité des niveaux critiques dans l'analyse technique.
}
\examples{
\dontrun{
library(zoo)
sr_data <- data.frame(
  Support = c(1.1, NA, NA, 1.4, 1.5),
  Resistance = c(2.1, 2.2, NA, NA, 2.5)
)
sr_fill_nas(sr_data)
}
}
